+++ CHANGELOG.md
@@ -12,6 +12,10 @@
 
 ## master
 
+#### :rocket: New Feature
+
+- Extend signature help to work on constructor payloads as well. Can be turned off if wanted through settings.
+
 ## 1.48.0
 
 #### :bug: Bug Fix